fix：文档上传

+++ server/askonce/components/dto/dto_gpt/response.go
@@ -41,6 +41,6 @@ type EmbeddingResp struct {
 
 type EmbeddingData struct {
 	Object    string    `json:"object"`
-	Embedding []float64 `json:"embedding"`
+	Embedding []float32 `json:"embedding"`
 	Index     int       `json:"index"`
 }
